requirements.txt
pre-commit
spotipy
segno
pillow
